+++ test/DocTests/Test/Predicates.hs
@@ -374,567 +374,567 @@ test = do
 {-# LINE 569 "src/Test/Predicates.hs" #-}
       (accept (caseInsensitive gt "NOTHING") "everything")
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:596: "
-{-# LINE 596 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 596 "src/Test/Predicates.hs" #-}
-      (accept (matchesRegex "x{2,5}y?") "xxxy")
-  [ExpectedLine [LineChunk "True"]]
  DocTest.printPrefix "Test.Predicates:598: "
 {-# LINE 598 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 598 "src/Test/Predicates.hs" #-}
-      (accept (matchesRegex "x{2,5}y?") "xyy")
-  [ExpectedLine [LineChunk "False"]]
+      (accept (matchesRegex "x{2,5}y?") "xxxy")
+  [ExpectedLine [LineChunk "True"]]
  DocTest.printPrefix "Test.Predicates:600: "
 {-# LINE 600 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 600 "src/Test/Predicates.hs" #-}
+      (accept (matchesRegex "x{2,5}y?") "xyy")
+  [ExpectedLine [LineChunk "False"]]
+ DocTest.printPrefix "Test.Predicates:602: "
+{-# LINE 602 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 602 "src/Test/Predicates.hs" #-}
       (accept (matchesRegex "x{2,5}y?") "wxxxyz")
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:630: "
-{-# LINE 630 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 630 "src/Test/Predicates.hs" #-}
-      (accept (matchesCaseInsensitiveRegex "x{2,5}y?") "XXXY")
-  [ExpectedLine [LineChunk "True"]]
  DocTest.printPrefix "Test.Predicates:632: "
 {-# LINE 632 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 632 "src/Test/Predicates.hs" #-}
-      (accept (matchesCaseInsensitiveRegex "x{2,5}y?") "XYY")
-  [ExpectedLine [LineChunk "False"]]
+      (accept (matchesCaseInsensitiveRegex "x{2,5}y?") "XXXY")
+  [ExpectedLine [LineChunk "True"]]
  DocTest.printPrefix "Test.Predicates:634: "
 {-# LINE 634 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 634 "src/Test/Predicates.hs" #-}
+      (accept (matchesCaseInsensitiveRegex "x{2,5}y?") "XYY")
+  [ExpectedLine [LineChunk "False"]]
+ DocTest.printPrefix "Test.Predicates:636: "
+{-# LINE 636 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 636 "src/Test/Predicates.hs" #-}
       (accept (matchesCaseInsensitiveRegex "x{2,5}y?") "WXXXYZ")
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:670: "
-{-# LINE 670 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 670 "src/Test/Predicates.hs" #-}
-      (accept (containsRegex "x{2,5}y?") "xxxy")
-  [ExpectedLine [LineChunk "True"]]
  DocTest.printPrefix "Test.Predicates:672: "
 {-# LINE 672 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 672 "src/Test/Predicates.hs" #-}
-      (accept (containsRegex "x{2,5}y?") "xyy")
-  [ExpectedLine [LineChunk "False"]]
+      (accept (containsRegex "x{2,5}y?") "xxxy")
+  [ExpectedLine [LineChunk "True"]]
  DocTest.printPrefix "Test.Predicates:674: "
 {-# LINE 674 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 674 "src/Test/Predicates.hs" #-}
-      (accept (containsRegex "x{2,5}y?") "wxxxyz")
-  [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:698: "
-{-# LINE 698 "src/Test/Predicates.hs" #-}
+      (accept (containsRegex "x{2,5}y?") "xyy")
+  [ExpectedLine [LineChunk "False"]]
+ DocTest.printPrefix "Test.Predicates:676: "
+{-# LINE 676 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 698 "src/Test/Predicates.hs" #-}
-      (accept (containsCaseInsensitiveRegex "x{2,5}y?") "XXXY")
+{-# LINE 676 "src/Test/Predicates.hs" #-}
+      (accept (containsRegex "x{2,5}y?") "wxxxyz")
   [ExpectedLine [LineChunk "True"]]
  DocTest.printPrefix "Test.Predicates:700: "
 {-# LINE 700 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 700 "src/Test/Predicates.hs" #-}
-      (accept (containsCaseInsensitiveRegex "x{2,5}y?") "XYY")
-  [ExpectedLine [LineChunk "False"]]
+      (accept (containsCaseInsensitiveRegex "x{2,5}y?") "XXXY")
+  [ExpectedLine [LineChunk "True"]]
  DocTest.printPrefix "Test.Predicates:702: "
 {-# LINE 702 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 702 "src/Test/Predicates.hs" #-}
+      (accept (containsCaseInsensitiveRegex "x{2,5}y?") "XYY")
+  [ExpectedLine [LineChunk "False"]]
+ DocTest.printPrefix "Test.Predicates:704: "
+{-# LINE 704 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 704 "src/Test/Predicates.hs" #-}
       (accept (containsCaseInsensitiveRegex "x{2,5}y?") "WXXXYZ")
   [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:726: "
-{-# LINE 726 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 726 "src/Test/Predicates.hs" #-}
-      (accept isEmpty ([] :: [Int]))
-  [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:728: "
-{-# LINE 728 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 728 "src/Test/Predicates.hs" #-}
-      (accept isEmpty [1, 2, 3])
-  [ExpectedLine [LineChunk "False"]]
  DocTest.printPrefix "Test.Predicates:730: "
 {-# LINE 730 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 730 "src/Test/Predicates.hs" #-}
-      (accept isEmpty "")
+      (accept isEmpty ([] :: [Int]))
   [ExpectedLine [LineChunk "True"]]
  DocTest.printPrefix "Test.Predicates:732: "
 {-# LINE 732 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 732 "src/Test/Predicates.hs" #-}
+      (accept isEmpty [1, 2, 3])
+  [ExpectedLine [LineChunk "False"]]
+ DocTest.printPrefix "Test.Predicates:734: "
+{-# LINE 734 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 734 "src/Test/Predicates.hs" #-}
+      (accept isEmpty "")
+  [ExpectedLine [LineChunk "True"]]
+ DocTest.printPrefix "Test.Predicates:736: "
+{-# LINE 736 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 736 "src/Test/Predicates.hs" #-}
       (accept isEmpty "gas tank")
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:745: "
-{-# LINE 745 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 745 "src/Test/Predicates.hs" #-}
-      (accept nonEmpty ([] :: [Int]))
-  [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:747: "
-{-# LINE 747 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 747 "src/Test/Predicates.hs" #-}
-      (accept nonEmpty [1, 2, 3])
-  [ExpectedLine [LineChunk "True"]]
  DocTest.printPrefix "Test.Predicates:749: "
 {-# LINE 749 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 749 "src/Test/Predicates.hs" #-}
-      (accept nonEmpty "")
+      (accept nonEmpty ([] :: [Int]))
   [ExpectedLine [LineChunk "False"]]
  DocTest.printPrefix "Test.Predicates:751: "
 {-# LINE 751 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 751 "src/Test/Predicates.hs" #-}
+      (accept nonEmpty [1, 2, 3])
+  [ExpectedLine [LineChunk "True"]]
+ DocTest.printPrefix "Test.Predicates:753: "
+{-# LINE 753 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 753 "src/Test/Predicates.hs" #-}
+      (accept nonEmpty "")
+  [ExpectedLine [LineChunk "False"]]
+ DocTest.printPrefix "Test.Predicates:755: "
+{-# LINE 755 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 755 "src/Test/Predicates.hs" #-}
       (accept nonEmpty "gas tank")
   [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:759: "
-{-# LINE 759 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:763: "
+{-# LINE 763 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 759 "src/Test/Predicates.hs" #-}
+{-# LINE 763 "src/Test/Predicates.hs" #-}
       (accept (sizeIs (lt 3)) ['a' .. 'f'])
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:761: "
-{-# LINE 761 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:765: "
+{-# LINE 765 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 761 "src/Test/Predicates.hs" #-}
+{-# LINE 765 "src/Test/Predicates.hs" #-}
       (accept (sizeIs (lt 3)) ['a' .. 'b'])
   [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:782: "
-{-# LINE 782 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 782 "src/Test/Predicates.hs" #-}
-      (accept (elemsAre [lt 3, lt 4, lt 5]) [2, 3, 4])
-  [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:784: "
-{-# LINE 784 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 784 "src/Test/Predicates.hs" #-}
-      (accept (elemsAre [lt 3, lt 4, lt 5]) [2, 3, 4, 5])
-  [ExpectedLine [LineChunk "False"]]
  DocTest.printPrefix "Test.Predicates:786: "
 {-# LINE 786 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 786 "src/Test/Predicates.hs" #-}
+      (accept (elemsAre [lt 3, lt 4, lt 5]) [2, 3, 4])
+  [ExpectedLine [LineChunk "True"]]
+ DocTest.printPrefix "Test.Predicates:788: "
+{-# LINE 788 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 788 "src/Test/Predicates.hs" #-}
+      (accept (elemsAre [lt 3, lt 4, lt 5]) [2, 3, 4, 5])
+  [ExpectedLine [LineChunk "False"]]
+ DocTest.printPrefix "Test.Predicates:790: "
+{-# LINE 790 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 790 "src/Test/Predicates.hs" #-}
       (accept (elemsAre [lt 3, lt 4, lt 5]) [2, 10, 4])
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:818: "
-{-# LINE 818 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 818 "src/Test/Predicates.hs" #-}
-      (accept (unorderedElemsAre [eq 1, eq 2, eq 3]) [1, 2, 3])
-  [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:820: "
-{-# LINE 820 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 820 "src/Test/Predicates.hs" #-}
-      (accept (unorderedElemsAre [eq 1, eq 2, eq 3]) [2, 3, 1])
-  [ExpectedLine [LineChunk "True"]]
  DocTest.printPrefix "Test.Predicates:822: "
 {-# LINE 822 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 822 "src/Test/Predicates.hs" #-}
-      (accept (unorderedElemsAre [eq 1, eq 2, eq 3]) [1, 2, 3, 4])
-  [ExpectedLine [LineChunk "False"]]
+      (accept (unorderedElemsAre [eq 1, eq 2, eq 3]) [1, 2, 3])
+  [ExpectedLine [LineChunk "True"]]
  DocTest.printPrefix "Test.Predicates:824: "
 {-# LINE 824 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 824 "src/Test/Predicates.hs" #-}
+      (accept (unorderedElemsAre [eq 1, eq 2, eq 3]) [2, 3, 1])
+  [ExpectedLine [LineChunk "True"]]
+ DocTest.printPrefix "Test.Predicates:826: "
+{-# LINE 826 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 826 "src/Test/Predicates.hs" #-}
+      (accept (unorderedElemsAre [eq 1, eq 2, eq 3]) [1, 2, 3, 4])
+  [ExpectedLine [LineChunk "False"]]
+ DocTest.printPrefix "Test.Predicates:828: "
+{-# LINE 828 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 828 "src/Test/Predicates.hs" #-}
       (accept (unorderedElemsAre [eq 1, eq 2, eq 3]) [1, 3])
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:871: "
-{-# LINE 871 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 871 "src/Test/Predicates.hs" #-}
-      (accept (each (gt 5)) [4, 5, 6])
-  [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:873: "
-{-# LINE 873 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 873 "src/Test/Predicates.hs" #-}
-      (accept (each (gt 5)) [6, 7, 8])
-  [ExpectedLine [LineChunk "True"]]
  DocTest.printPrefix "Test.Predicates:875: "
 {-# LINE 875 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 875 "src/Test/Predicates.hs" #-}
-      (accept (each (gt 5)) [])
-  [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:900: "
-{-# LINE 900 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 900 "src/Test/Predicates.hs" #-}
-      (accept (contains (gt 5)) [3, 4, 5])
+      (accept (each (gt 5)) [4, 5, 6])
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:902: "
-{-# LINE 902 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:877: "
+{-# LINE 877 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 902 "src/Test/Predicates.hs" #-}
-      (accept (contains (gt 5)) [4, 5, 6])
+{-# LINE 877 "src/Test/Predicates.hs" #-}
+      (accept (each (gt 5)) [6, 7, 8])
+  [ExpectedLine [LineChunk "True"]]
+ DocTest.printPrefix "Test.Predicates:879: "
+{-# LINE 879 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 879 "src/Test/Predicates.hs" #-}
+      (accept (each (gt 5)) [])
   [ExpectedLine [LineChunk "True"]]
  DocTest.printPrefix "Test.Predicates:904: "
 {-# LINE 904 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 904 "src/Test/Predicates.hs" #-}
-      (accept (contains (gt 5)) [])
+      (accept (contains (gt 5)) [3, 4, 5])
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:912: "
-{-# LINE 912 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:906: "
+{-# LINE 906 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 912 "src/Test/Predicates.hs" #-}
-      (accept (containsAll [eq "foo", eq "bar"]) ["bar", "foo"])
+{-# LINE 906 "src/Test/Predicates.hs" #-}
+      (accept (contains (gt 5)) [4, 5, 6])
   [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:914: "
-{-# LINE 914 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:908: "
+{-# LINE 908 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 914 "src/Test/Predicates.hs" #-}
-      (accept (containsAll [eq "foo", eq "bar"]) ["foo"])
+{-# LINE 908 "src/Test/Predicates.hs" #-}
+      (accept (contains (gt 5)) [])
   [ExpectedLine [LineChunk "False"]]
  DocTest.printPrefix "Test.Predicates:916: "
 {-# LINE 916 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 916 "src/Test/Predicates.hs" #-}
+      (accept (containsAll [eq "foo", eq "bar"]) ["bar", "foo"])
+  [ExpectedLine [LineChunk "True"]]
+ DocTest.printPrefix "Test.Predicates:918: "
+{-# LINE 918 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 918 "src/Test/Predicates.hs" #-}
+      (accept (containsAll [eq "foo", eq "bar"]) ["foo"])
+  [ExpectedLine [LineChunk "False"]]
+ DocTest.printPrefix "Test.Predicates:920: "
+{-# LINE 920 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 920 "src/Test/Predicates.hs" #-}
       (accept (containsAll [eq "foo", eq "bar"]) ["foo", "bar", "qux"])
   [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:924: "
-{-# LINE 924 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:928: "
+{-# LINE 928 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 924 "src/Test/Predicates.hs" #-}
+{-# LINE 928 "src/Test/Predicates.hs" #-}
       (accept (containsAll [startsWith "f", endsWith "o"]) ["foo"])
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:926: "
-{-# LINE 926 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:930: "
+{-# LINE 930 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 926 "src/Test/Predicates.hs" #-}
+{-# LINE 930 "src/Test/Predicates.hs" #-}
       (accept (contains (startsWith "f") `andP` contains (endsWith "o")) ["foo"])
-  [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:948: "
-{-# LINE 948 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 948 "src/Test/Predicates.hs" #-}
-      (accept (containsOnly [eq "foo", eq "bar"]) ["foo"])
-  [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:950: "
-{-# LINE 950 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 950 "src/Test/Predicates.hs" #-}
-      (accept (containsOnly [eq "foo", eq "bar"]) ["foo", "bar"])
   [ExpectedLine [LineChunk "True"]]
  DocTest.printPrefix "Test.Predicates:952: "
 {-# LINE 952 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 952 "src/Test/Predicates.hs" #-}
+      (accept (containsOnly [eq "foo", eq "bar"]) ["foo"])
+  [ExpectedLine [LineChunk "True"]]
+ DocTest.printPrefix "Test.Predicates:954: "
+{-# LINE 954 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 954 "src/Test/Predicates.hs" #-}
+      (accept (containsOnly [eq "foo", eq "bar"]) ["foo", "bar"])
+  [ExpectedLine [LineChunk "True"]]
+ DocTest.printPrefix "Test.Predicates:956: "
+{-# LINE 956 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 956 "src/Test/Predicates.hs" #-}
       (accept (containsOnly [eq "foo", eq "bar"]) ["foo", "qux"])
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:959: "
-{-# LINE 959 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:963: "
+{-# LINE 963 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 959 "src/Test/Predicates.hs" #-}
+{-# LINE 963 "src/Test/Predicates.hs" #-}
       (accept (containsOnly [eq "foo", eq "bar"]) ["foo", "foo"])
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:961: "
-{-# LINE 961 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:965: "
+{-# LINE 965 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 961 "src/Test/Predicates.hs" #-}
+{-# LINE 965 "src/Test/Predicates.hs" #-}
       (accept (each (eq "foo" `orP` eq "bar")) ["foo", "foo"])
   [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:987: "
-{-# LINE 987 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:991: "
+{-# LINE 991 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 987 "src/Test/Predicates.hs" #-}
+{-# LINE 991 "src/Test/Predicates.hs" #-}
       (accept (keys (each (eq "foo"))) [("foo", 5)])
   [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:990: "
-{-# LINE 990 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:994: "
+{-# LINE 994 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 990 "src/Test/Predicates.hs" #-}
+{-# LINE 994 "src/Test/Predicates.hs" #-}
       (accept (keys (each (eq "foo"))) [("foo", 5), ("bar", 6)])
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:1006: "
-{-# LINE 1006 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1010: "
+{-# LINE 1010 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1006 "src/Test/Predicates.hs" #-}
+{-# LINE 1010 "src/Test/Predicates.hs" #-}
       (accept (values (each (eq 5))) [("foo", 5), ("bar", 5)])
   [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:1009: "
-{-# LINE 1009 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1013: "
+{-# LINE 1013 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1009 "src/Test/Predicates.hs" #-}
+{-# LINE 1013 "src/Test/Predicates.hs" #-}
       (accept (values (each (eq 5))) [("foo", 5), ("bar", 6)])
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:1027: "
-{-# LINE 1027 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1031: "
+{-# LINE 1031 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1027 "src/Test/Predicates.hs" #-}
+{-# LINE 1031 "src/Test/Predicates.hs" #-}
       (accept (eq 1.0) (sum (replicate 100 0.01)))
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:1034: "
-{-# LINE 1034 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1038: "
+{-# LINE 1038 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1034 "src/Test/Predicates.hs" #-}
+{-# LINE 1038 "src/Test/Predicates.hs" #-}
       (accept (approxEq 1.0) (sum (replicate 100 0.01)))
   [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:1036: "
-{-# LINE 1036 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1040: "
+{-# LINE 1040 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1036 "src/Test/Predicates.hs" #-}
+{-# LINE 1040 "src/Test/Predicates.hs" #-}
       (accept (approxEq 1.0) (sum (replicate 100 0.009999)))
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:1051: "
-{-# LINE 1051 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1055: "
+{-# LINE 1055 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1051 "src/Test/Predicates.hs" #-}
+{-# LINE 1055 "src/Test/Predicates.hs" #-}
       (accept positive 1)
   [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:1054: "
-{-# LINE 1054 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1058: "
+{-# LINE 1058 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1054 "src/Test/Predicates.hs" #-}
+{-# LINE 1058 "src/Test/Predicates.hs" #-}
       (accept positive 0)
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:1057: "
-{-# LINE 1057 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1061: "
+{-# LINE 1061 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1057 "src/Test/Predicates.hs" #-}
+{-# LINE 1061 "src/Test/Predicates.hs" #-}
       (accept positive (-1))
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:1075: "
-{-# LINE 1075 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1079: "
+{-# LINE 1079 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1075 "src/Test/Predicates.hs" #-}
+{-# LINE 1079 "src/Test/Predicates.hs" #-}
       (accept negative 1)
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:1078: "
-{-# LINE 1078 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1082: "
+{-# LINE 1082 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1078 "src/Test/Predicates.hs" #-}
+{-# LINE 1082 "src/Test/Predicates.hs" #-}
       (accept negative 0)
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:1081: "
-{-# LINE 1081 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1085: "
+{-# LINE 1085 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1081 "src/Test/Predicates.hs" #-}
+{-# LINE 1085 "src/Test/Predicates.hs" #-}
       (accept negative (-1))
   [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:1100: "
-{-# LINE 1100 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1104: "
+{-# LINE 1104 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1100 "src/Test/Predicates.hs" #-}
+{-# LINE 1104 "src/Test/Predicates.hs" #-}
       (accept nonPositive 1)
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:1103: "
-{-# LINE 1103 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1107: "
+{-# LINE 1107 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1103 "src/Test/Predicates.hs" #-}
+{-# LINE 1107 "src/Test/Predicates.hs" #-}
       (accept nonPositive 0)
   [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:1106: "
-{-# LINE 1106 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1110: "
+{-# LINE 1110 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1106 "src/Test/Predicates.hs" #-}
+{-# LINE 1110 "src/Test/Predicates.hs" #-}
       (accept nonPositive (-1))
   [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:1114: "
-{-# LINE 1114 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1118: "
+{-# LINE 1118 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1114 "src/Test/Predicates.hs" #-}
+{-# LINE 1118 "src/Test/Predicates.hs" #-}
       (accept nonNegative 1)
   [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:1117: "
-{-# LINE 1117 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1121: "
+{-# LINE 1121 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1117 "src/Test/Predicates.hs" #-}
+{-# LINE 1121 "src/Test/Predicates.hs" #-}
       (accept nonNegative 0)
   [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:1120: "
-{-# LINE 1120 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1124: "
+{-# LINE 1124 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1120 "src/Test/Predicates.hs" #-}
+{-# LINE 1124 "src/Test/Predicates.hs" #-}
       (accept nonNegative (-1))
-  [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:1127: "
-{-# LINE 1127 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 1127 "src/Test/Predicates.hs" #-}
-      (accept finite 1.0)
-  [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:1129: "
-{-# LINE 1129 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 1129 "src/Test/Predicates.hs" #-}
-      (accept finite (0 / 0))
   [ExpectedLine [LineChunk "False"]]
  DocTest.printPrefix "Test.Predicates:1131: "
 {-# LINE 1131 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 1131 "src/Test/Predicates.hs" #-}
+      (accept finite 1.0)
+  [ExpectedLine [LineChunk "True"]]
+ DocTest.printPrefix "Test.Predicates:1133: "
+{-# LINE 1133 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 1133 "src/Test/Predicates.hs" #-}
+      (accept finite (0 / 0))
+  [ExpectedLine [LineChunk "False"]]
+ DocTest.printPrefix "Test.Predicates:1135: "
+{-# LINE 1135 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 1135 "src/Test/Predicates.hs" #-}
       (accept finite (1 / 0))
-  [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:1149: "
-{-# LINE 1149 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 1149 "src/Test/Predicates.hs" #-}
-      (accept infinite 1.0)
-  [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:1151: "
-{-# LINE 1151 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 1151 "src/Test/Predicates.hs" #-}
-      (accept infinite (0 / 0))
   [ExpectedLine [LineChunk "False"]]
  DocTest.printPrefix "Test.Predicates:1153: "
 {-# LINE 1153 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 1153 "src/Test/Predicates.hs" #-}
-      (accept infinite (1 / 0))
-  [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:1169: "
-{-# LINE 1169 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 1169 "src/Test/Predicates.hs" #-}
-      (accept nAn 1.0)
+      (accept infinite 1.0)
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:1171: "
-{-# LINE 1171 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1155: "
+{-# LINE 1155 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1171 "src/Test/Predicates.hs" #-}
-      (accept nAn (0 / 0))
+{-# LINE 1155 "src/Test/Predicates.hs" #-}
+      (accept infinite (0 / 0))
+  [ExpectedLine [LineChunk "False"]]
+ DocTest.printPrefix "Test.Predicates:1157: "
+{-# LINE 1157 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 1157 "src/Test/Predicates.hs" #-}
+      (accept infinite (1 / 0))
   [ExpectedLine [LineChunk "True"]]
  DocTest.printPrefix "Test.Predicates:1173: "
 {-# LINE 1173 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 1173 "src/Test/Predicates.hs" #-}
+      (accept nAn 1.0)
+  [ExpectedLine [LineChunk "False"]]
+ DocTest.printPrefix "Test.Predicates:1175: "
+{-# LINE 1175 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 1175 "src/Test/Predicates.hs" #-}
+      (accept nAn (0 / 0))
+  [ExpectedLine [LineChunk "True"]]
+ DocTest.printPrefix "Test.Predicates:1177: "
+{-# LINE 1177 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 1177 "src/Test/Predicates.hs" #-}
       (accept nAn (1 / 0))
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:1192: "
-{-# LINE 1192 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1196: "
+{-# LINE 1196 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1192 "src/Test/Predicates.hs" #-}
+{-# LINE 1196 "src/Test/Predicates.hs" #-}
       (accept (is even) 3)
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:1194: "
-{-# LINE 1194 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1198: "
+{-# LINE 1198 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1194 "src/Test/Predicates.hs" #-}
+{-# LINE 1198 "src/Test/Predicates.hs" #-}
       (accept (is even) 4)
   [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:1211: "
-{-# LINE 1211 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1215: "
+{-# LINE 1215 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1211 "src/Test/Predicates.hs" #-}
+{-# LINE 1215 "src/Test/Predicates.hs" #-}
       (accept $(qIs [| even |]) 3)
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:1213: "
-{-# LINE 1213 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1217: "
+{-# LINE 1217 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1213 "src/Test/Predicates.hs" #-}
+{-# LINE 1217 "src/Test/Predicates.hs" #-}
       (accept $(qIs [| even |]) 4)
   [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:1216: "
-{-# LINE 1216 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1220: "
+{-# LINE 1220 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1216 "src/Test/Predicates.hs" #-}
+{-# LINE 1220 "src/Test/Predicates.hs" #-}
       (show $(qIs [| even |]))
   [ExpectedLine [LineChunk "\"even\""]]
- DocTest.printPrefix "Test.Predicates:1236: "
-{-# LINE 1236 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 1236 "src/Test/Predicates.hs" #-}
-      (accept (with abs (gt 5)) (-6))
-  [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:1238: "
-{-# LINE 1238 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 1238 "src/Test/Predicates.hs" #-}
-      (accept (with abs (gt 5)) (-5))
-  [ExpectedLine [LineChunk "False"]]
  DocTest.printPrefix "Test.Predicates:1240: "
 {-# LINE 1240 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 1240 "src/Test/Predicates.hs" #-}
-      (accept (with reverse (eq "olleh")) "hello")
+      (accept (with abs (gt 5)) (-6))
   [ExpectedLine [LineChunk "True"]]
  DocTest.printPrefix "Test.Predicates:1242: "
 {-# LINE 1242 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 1242 "src/Test/Predicates.hs" #-}
-      (accept (with reverse (eq "olleh")) "goodbye")
+      (accept (with abs (gt 5)) (-5))
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:1268: "
-{-# LINE 1268 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1244: "
+{-# LINE 1244 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1268 "src/Test/Predicates.hs" #-}
-      (accept ($(qWith [| abs |]) (gt 5)) (-6))
+{-# LINE 1244 "src/Test/Predicates.hs" #-}
+      (accept (with reverse (eq "olleh")) "hello")
   [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:1270: "
-{-# LINE 1270 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1246: "
+{-# LINE 1246 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1270 "src/Test/Predicates.hs" #-}
-      (accept ($(qWith [| abs |]) (gt 5)) (-5))
+{-# LINE 1246 "src/Test/Predicates.hs" #-}
+      (accept (with reverse (eq "olleh")) "goodbye")
   [ExpectedLine [LineChunk "False"]]
  DocTest.printPrefix "Test.Predicates:1272: "
 {-# LINE 1272 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 1272 "src/Test/Predicates.hs" #-}
-      (accept ($(qWith [| reverse |]) (eq "olleh")) "hello")
+      (accept ($(qWith [| abs |]) (gt 5)) (-6))
   [ExpectedLine [LineChunk "True"]]
  DocTest.printPrefix "Test.Predicates:1274: "
 {-# LINE 1274 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 1274 "src/Test/Predicates.hs" #-}
+      (accept ($(qWith [| abs |]) (gt 5)) (-5))
+  [ExpectedLine [LineChunk "False"]]
+ DocTest.printPrefix "Test.Predicates:1276: "
+{-# LINE 1276 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 1276 "src/Test/Predicates.hs" #-}
+      (accept ($(qWith [| reverse |]) (eq "olleh")) "hello")
+  [ExpectedLine [LineChunk "True"]]
+ DocTest.printPrefix "Test.Predicates:1278: "
+{-# LINE 1278 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 1278 "src/Test/Predicates.hs" #-}
       (accept ($(qWith [| reverse |]) (eq "olleh")) "goodbye")
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:1277: "
-{-# LINE 1277 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1281: "
+{-# LINE 1281 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1277 "src/Test/Predicates.hs" #-}
+{-# LINE 1281 "src/Test/Predicates.hs" #-}
       (show ($(qWith [| abs |]) (gt 5)))
   [ExpectedLine [LineChunk "\"abs: > 5\""]]
- DocTest.printPrefix "Test.Predicates:1296: "
-{-# LINE 1296 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 1296 "src/Test/Predicates.hs" #-}
-      (accept $(qMatch [p| Just (Left _) |]) Nothing)
-  [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:1298: "
-{-# LINE 1298 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 1298 "src/Test/Predicates.hs" #-}
-      (accept $(qMatch [p| Just (Left _) |]) (Just (Left 5)))
-  [ExpectedLine [LineChunk "True"]]
  DocTest.printPrefix "Test.Predicates:1300: "
 {-# LINE 1300 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 1300 "src/Test/Predicates.hs" #-}
+      (accept $(qMatch [p| Just (Left _) |]) Nothing)
+  [ExpectedLine [LineChunk "False"]]
+ DocTest.printPrefix "Test.Predicates:1302: "
+{-# LINE 1302 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 1302 "src/Test/Predicates.hs" #-}
+      (accept $(qMatch [p| Just (Left _) |]) (Just (Left 5)))
+  [ExpectedLine [LineChunk "True"]]
+ DocTest.printPrefix "Test.Predicates:1304: "
+{-# LINE 1304 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 1304 "src/Test/Predicates.hs" #-}
       (accept $(qMatch [p| Just (Left _) |]) (Just (Right 5)))
   [ExpectedLine [LineChunk "False"]]
- DocTest.printPrefix "Test.Predicates:1303: "
-{-# LINE 1303 "src/Test/Predicates.hs" #-}
+ DocTest.printPrefix "Test.Predicates:1307: "
+{-# LINE 1307 "src/Test/Predicates.hs" #-}
  DocTest.example
-{-# LINE 1303 "src/Test/Predicates.hs" #-}
+{-# LINE 1307 "src/Test/Predicates.hs" #-}
       (show $(qMatch [p| Just (Left _) |]))
   [ExpectedLine [LineChunk "\"Just (Left _)\""]]
- DocTest.printPrefix "Test.Predicates:1325: "
-{-# LINE 1325 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 1325 "src/Test/Predicates.hs" #-}
-      (accept (typed @String anything) "foo")
-  [ExpectedLine [LineChunk "True"]]
- DocTest.printPrefix "Test.Predicates:1327: "
-{-# LINE 1327 "src/Test/Predicates.hs" #-}
- DocTest.example
-{-# LINE 1327 "src/Test/Predicates.hs" #-}
-      (accept (typed @String (sizeIs (gt 5))) "foo")
-  [ExpectedLine [LineChunk "False"]]
  DocTest.printPrefix "Test.Predicates:1329: "
 {-# LINE 1329 "src/Test/Predicates.hs" #-}
  DocTest.example
 {-# LINE 1329 "src/Test/Predicates.hs" #-}
+      (accept (typed @String anything) "foo")
+  [ExpectedLine [LineChunk "True"]]
+ DocTest.printPrefix "Test.Predicates:1331: "
+{-# LINE 1331 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 1331 "src/Test/Predicates.hs" #-}
+      (accept (typed @String (sizeIs (gt 5))) "foo")
+  [ExpectedLine [LineChunk "False"]]
+ DocTest.printPrefix "Test.Predicates:1333: "
+{-# LINE 1333 "src/Test/Predicates.hs" #-}
+ DocTest.example
+{-# LINE 1333 "src/Test/Predicates.hs" #-}
       (accept (typed @String anything) (42 :: Int))
   [ExpectedLine [LineChunk "False"]]